requirements.txt
requests==2.18.4
pyhcl==0.3.9
te2-sdk==0.1.5
python-consul==0.7.2
click==6.7
jinja2==2.10